+++ dsh-api/src/secret.rs
@@ -20,22 +20,26 @@
 //! [`DshApiClient`] methods that add extra capabilities but do not directly call the
 //! DSH resource management API. These derived methods depend on the API methods for this.
 //!
-//! * [`secrets_with_dependants() -> [id, [application|app]]`](DshApiClient::secrets_with_dependants)
-//! * [`secrets_with_dependant_applications() -> [id, [application]]`](DshApiClient::secrets_with_dependant_applications)
-//! * [`secrets_with_dependant_apps() -> [id, [app]]`](DshApiClient::secrets_with_dependant_apps)
+//! * [`secret_dependants(id) -> Vec<Dependant>`](DshApiClient::secret_dependants)
+//! * [`secrets_with_dependants() -> Vec<(String, Vec<Dependant)>`](DshApiClient::secrets_with_dependants)
+//! * [`secrets_with_dependant_applications() -> Vec<(String, Vec<DependantApplication)>`](DshApiClient::secrets_with_dependant_applications)
+//! * [`secrets_with_dependant_apps() -> Vec<(String, Vec<DependantApp)>`](DshApiClient::secrets_with_dependant_apps)
+//! * [`secrets_with_dependant_proxies() -> Vec<(String, Vec<DependantProxy)>`](DshApiClient::secrets_with_dependant_proxies)
 
 use crate::app::{app_resources, apps_that_use_secret};
 use crate::application_types::{ApplicationValues, EnvVarInjection};
 use crate::dsh_api_client::DshApiClient;
+use crate::proxy::proxies_that_use_secret;
 #[allow(unused_imports)]
 use crate::types::{AllocationStatus, Empty, Secret};
 use crate::types::{AppCatalogApp, AppCatalogAppResourcesValue, Application};
 #[allow(unused_imports)]
 use crate::DshApiError;
-use crate::{Dependant, DependantApp, DependantApplication, DshApiResult};
+use crate::{Dependant, DependantApp, DependantApplication, DependantProxy, DshApiResult};
 use futures::try_join;
 use itertools::Itertools;
 use serde::{Deserialize, Serialize};
+use std::borrow::Cow;
 use std::collections::HashMap;
 use std::fmt::{Display, Formatter};
 
@@ -67,34 +71,69 @@ impl Display for SecretInjection {
 /// [`DshApiClient`] methods that add extra capabilities but do not directly call the
 /// DSH resource management API. These derived methods depend on the API methods for this.
 ///
-/// * [`secrets_with_dependants() -> [id, [application|app]]`](DshApiClient::secrets_with_dependants)
-/// * [`secrets_with_dependant_applications() -> [id, [application]]`](DshApiClient::secrets_with_dependant_applications)
-/// * [`secrets_with_dependant_apps() -> [id, [app]]`](DshApiClient::secrets_with_dependant_apps)
+/// * [`secret_dependants(id) -> Vec<Dependant>`](DshApiClient::secret_dependants)
+/// * [`secrets_with_dependants() -> Vec<(String, Vec<Dependant)>`](DshApiClient::secrets_with_dependants)
+/// * [`secrets_with_dependant_applications() -> Vec<(String, Vec<DependantApplication)>`](DshApiClient::secrets_with_dependant_applications)
+/// * [`secrets_with_dependant_apps() -> Vec<(String, Vec<DependantApp)>`](DshApiClient::secrets_with_dependant_apps)
+/// * [`secrets_with_dependant_proxies() -> Vec<(String, Vec<DependantProxy)>`](DshApiClient::secrets_with_dependant_proxies)
 impl DshApiClient {
-  /// # Returns all secrets with dependant applications and apps
+  /// # Returns all secrets with dependant applications, apps and proxies
   ///
-  /// Returns a sorted list of all secrets together with the applications and apps that use them.
+  /// # Parameters
+  /// * `secret_id` - Id of the secret to get the dependants for.
+  ///
+  /// Returns the applications, apps and proxies that use the provided secret.
+  pub async fn secret_dependants(&self, secret_id: &str) -> DshApiResult<Vec<Dependant<SecretInjection>>> {
+    let (applications, apps, proxies) = try_join!(self.get_application_configuration_map(), self.get_appcatalogapp_configuration_map(), self.proxies())?;
+    let mut dependants: Vec<Dependant<SecretInjection>> = vec![];
+    for application in secret_env_vars_from_applications(secret_id, &applications) {
+      dependants.push(Dependant::application(
+        application.id.to_string(),
+        application.application.instances,
+        application.values.iter().map(|env_var| SecretInjection::EnvVar(env_var.to_string())).collect_vec(),
+      ));
+    }
+    for (app_id, _, resource_ids) in apps_that_use_secret(secret_id, &apps) {
+      dependants.push(Dependant::app(
+        app_id.to_string(),
+        resource_ids.iter().map(|resource_id| resource_id.to_string()).collect_vec(),
+      ));
+    }
+    for (proxy_id, proxy) in proxies_that_use_secret(secret_id, &proxies) {
+      dependants.push(Dependant::proxy(proxy_id.to_string(), proxy.instances.get()));
+    }
+    Ok(dependants)
+  }
+
+  /// # Returns all secrets with dependant applications, apps and proxies
+  ///
+  /// Returns a sorted list of all secrets together with the applications, apps and proxies that
+  /// use them.
   pub async fn secrets_with_dependants(&self) -> DshApiResult<Vec<(String, Vec<Dependant<SecretInjection>>)>> {
-    let (secret_ids, applications, apps) = try_join!(
+    let (secret_ids, applications, apps, proxies) = try_join!(
       self.get_secret_ids(),
       self.get_application_configuration_map(),
-      self.get_appcatalogapp_configuration_map()
+      self.get_appcatalogapp_configuration_map(),
+      self.proxies()
     )?;
     let mut secrets = Vec::<(String, Vec<Dependant<SecretInjection>>)>::new();
     for secret_id in secret_ids {
       let mut dependants: Vec<Dependant<SecretInjection>> = vec![];
-      for application in secret_env_vars_from_applications(secret_id.as_str(), &applications) {
+      for application in secret_env_vars_from_applications(&secret_id, &applications) {
         dependants.push(Dependant::application(
           application.id.to_string(),
           application.application.instances,
           application.values.iter().map(|env_var| SecretInjection::EnvVar(env_var.to_string())).collect_vec(),
         ));
       }
-      for (app_id, _, resource_ids) in apps_that_use_secret(secret_id.as_str(), &apps) {
+      for (app_id, _, resource_ids) in apps_that_use_secret(&secret_id, &apps) {
         dependants.push(Dependant::app(
           app_id.to_string(),
           resource_ids.iter().map(|resource_id| resource_id.to_string()).collect_vec(),
         ));
+      }
+      for (proxy_id, proxy) in proxies_that_use_secret(&secret_id, &proxies) {
+        dependants.push(Dependant::proxy(proxy_id.to_string(), proxy.instances.get()));
       }
       secrets.push((secret_id, dependants));
     }
@@ -139,6 +178,22 @@ impl DshApiClient {
     }
     Ok(secrets)
   }
+
+  /// # Returns all secrets with dependant proxies
+  ///
+  /// Returns a sorted list of all secrets together with the proxies that use them.
+  pub async fn secrets_with_dependant_proxies(&self) -> DshApiResult<Vec<(String, Vec<DependantProxy>)>> {
+    let (secret_ids, proxies) = try_join!(self.get_secret_ids(), self.proxies())?;
+    let mut secrets = Vec::<(String, Vec<DependantProxy>)>::new();
+    for secret_id in secret_ids {
+      let mut dependant_proxies: Vec<DependantProxy> = vec![];
+      for (proxy_id, proxy) in proxies_that_use_secret(secret_id.as_str(), &proxies) {
+        dependant_proxies.push(DependantProxy::new(proxy_id.to_string(), proxy.instances.get()));
+      }
+      secrets.push((secret_id, dependant_proxies));
+    }
+    Ok(secrets)
+  }
 }
 
 /// # Get application environment variables referencing secret
@@ -146,8 +201,8 @@ impl DshApiClient {
 /// Get all environment variables from `application` referencing secret with `secret_id`.
 ///
 /// # Parameters
-/// * `secret_id` - id of the secret to look for
-/// * `application` - reference to the `Application`
+/// * `secret_id` - Id of the secret to look for.
+/// * `application` - Reference to the `Application`.
 ///
 /// # Returns
 /// * `Vec<EnvVarKey>` - list of all environment variables referencing secret `secret_id`
@@ -185,8 +240,8 @@ pub fn secret_env_vars_from_application<'a>(secret_id: &str, application: &'a Ap
 /// Applications are only included if they reference secret `secret_id` at least once.
 ///
 /// # Parameters
-/// * `secret_id` - id of the secret to look for
-/// * `applications` - hashmap containing id/application pairs
+/// * `secret_id` - Id of the secret to look for.
+/// * `applications` - Hashmap containing id/application pairs.
 ///
 /// # Returns
 /// `Vec<ApplicationTuple<EnvVarKey>>` - list of tuples containing:
@@ -211,15 +266,41 @@ pub fn secret_env_vars_from_applications<'a>(secret_id: &str, applications: &'a 
   application_tuples
 }
 
-/// # Checks if secret is a system secret
+/// # Checks if secret id is a system secret
+///
+/// Deprecated, use [is_system_id].
+#[deprecated]
 pub fn secret_is_system(secret_id: &str) -> bool {
   secret_id.contains('!')
+}
+
+/// # Checks if secret id is a system secret
+pub fn is_system_id(secret_id: &str) -> bool {
+  secret_id.contains('!')
+}
+
+/// # Converts secret id to secret name
+///
+/// Converts secret id to secret name, if not the same.
+///
+/// # Parameters
+/// `secret_id` - Secret id to be converted.
+///
+/// # Returns
+/// `Cow::Borrowed` - Secret id was already in the proper format.
+/// `Cow::Owned` - Secret id was not in the proper format.
+pub fn secret_id_to_secret_name(secret_id: &String) -> Cow<String> {
+  if is_system_id(secret_id) {
+    Cow::Owned(format!("system{}", secret_id.replace("!", "/")))
+  } else {
+    Cow::Borrowed(secret_id)
+  }
 }
 
 /// Get secret resources from `AppCatalogApp`
 ///
 /// # Parameters
-/// * `app` - app to get the secret resources from
+/// * `app` - App to get the secret resources from.
 ///
 /// # Returns
 /// Either `None` when the `app` does not have any secret resources,
@@ -238,7 +319,7 @@ pub fn secret_resources_from_app(app: &AppCatalogApp) -> Vec<(&str, &Secret)> {
 /// Get all environment variables from an `Application` that reference secrets.
 ///
 /// # Parameters
-/// * `application` - reference to the `Application`
+/// * `application` - Reference to the `Application`.
 ///
 /// # Returns
 /// `Vec<EnvInjection>` - list of tuples containing:
@@ -275,7 +356,7 @@ pub fn secrets_from_application(application: &Application) -> Vec<EnvVarInjectio
 /// Get all environment variables referencing secrets from all `Applications`
 ///
 /// # Parameters
-/// * `applications` - hashmap containing id/application pairs
+/// * `applications` - Hashmap containing id/application pairs.
 ///
 /// # Returns
 /// `Vec<ApplicationTuple<EnvInjection>>` - list of tuples containing:
@@ -303,8 +384,8 @@ pub fn secrets_from_applications(applications: &HashMap<String, Application>) ->
 /// Find apps that use any of a list of given secret
 ///
 /// # Parameters
-/// * `secrets` - ids of the secrets to look for
-/// * `apps` - hashmap of all apps
+/// * `secrets` - Ids of the secrets to look for.
+/// * `apps` - Hashmap of all apps.
 ///
 /// # Returns
 /// * `Vec<(app_id, app, resource_ids)>` - vector of apps that use the secret